[fix] The project elixir isn't always present
When running in a language server context, it's possible that the
version isn't set in the project. Similarly, there are other contexts
where you might want to use Quokka (like as a library).

The current version parsing relies on the project existing and its
elixir version being set. If the latter isn't true, then the current
code returns nil, which causes the regex match to fail.

This fixes the failure by defaulting to an empty string, which will
cause the system version to be used.